# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ServITech API",
        "description": "API documentation for ServITech",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/{version}/user/profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get the authenticated user profile",
                "description": "Get the authenticated User.",
                "operationId": "657d401dbfb80a9d121434dc2add79b9",
                "parameters": [
                    {
                        "name": "version",
                        "in": "path",
                        "description": "API version",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "v1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User information retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "$ref": "#/components/schemas/UserResource"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update the user basic information",
                "description": "Update the user basic information in storage.",
                "operationId": "550f603dfd27f45535e97d26cc87031f",
                "parameters": [
                    {
                        "name": "version",
                        "in": "path",
                        "description": "API version",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "v1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User information updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User information updated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "$ref": "#/components/schemas/UserResource"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/{version}/user/password": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update the user password",
                "description": "Update the user password in storage.",
                "operationId": "50c8fca358616ad7d4e016c24c51d1df",
                "parameters": [
                    {
                        "name": "version",
                        "in": "path",
                        "description": "API version",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "v1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User information updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User password updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UpdateUserRequest": {
                "required": [
                    "name",
                    "last_name"
                ],
                "properties": {
                    "name": {
                        "description": "Class UserController for handling user management.",
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    }
                },
                "type": "object"
            },
            "UpdatePasswordRequest": {
                "required": [
                    "old_password",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "old_password": {
                        "type": "string",
                        "format": "password",
                        "example": "oldpassword123"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "newpassword123"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "format": "password",
                        "example": "newpassword123"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "name": {
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}